+++ src/sequence/rainbow.rs
@@ -1,6 +1,6 @@
 // led_effects - A collection of LED effects on top of smart-leds.
 // Copyright (C) 2021 Guillaume Cugnet <guillaume@cugnet.eu>
-// Copyright (C) 2021 Jean-Philippe Cugnet <jean-philippe@cugnet.eu>
+// Copyright (C) 2021-2022 Jean-Philippe Cugnet <jean-philippe@cugnet.eu>
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -16,26 +16,34 @@
 
 use smart_leds::{hsv::Hsv, RGB8};
 
-use super::{OneParameterSequence, Sequence};
+use super::{ConfigWithMainColor, Sequence};
 
 /// A sequence in which the LEDs draw a rainbow.
 pub struct Rainbow<const N: usize> {
-    /// The first color of the rainbow.
-    first_color: Hsv,
+    /// The configuration.
+    config: RainbowConfig,
     /// The counter.
     counter: usize,
 }
 
-impl<const N: usize> Sequence<N> for Rainbow<N> {}
+/// The configuration for rainbow sequences.
+#[derive(Clone, Copy)]
+pub struct RainbowConfig {
+    /// The first color of the rainbow.
+    pub first_color: Hsv,
+    /// The color range.
+    pub range: u8,
+}
 
-impl<Color: Into<Hsv>, const N: usize> OneParameterSequence<Color, N>
-    for Rainbow<N>
-{
-    fn new(first_color: Color) -> Self {
-        Self {
-            first_color: first_color.into(),
-            counter: 0,
-        }
+impl<const N: usize> Sequence<N> for Rainbow<N> {
+    type Config = RainbowConfig;
+
+    fn new(config: Self::Config) -> Self {
+        Self { config, counter: 0 }
+    }
+
+    fn config(&self) -> Self::Config {
+        self.config
     }
 }
 
@@ -45,13 +53,24 @@ impl<const N: usize> Iterator for Rainbow<N> {
     fn next(&mut self) -> Option<Self::Item> {
         if self.counter < N {
             let color = Hsv {
-                hue: self.first_color.hue + (self.counter * (255 / N)) as u8,
-                ..self.first_color // sat: self.first_color.sat, val: self.first_color.val
+                hue: self.config.first_color.hue
+                    + (self.counter * (self.config.range as usize / N)) as u8,
+                ..self.config.first_color
             };
             self.counter += 1;
             Some(color.into())
         } else {
             None
         }
+    }
+}
+
+impl ConfigWithMainColor for RainbowConfig {
+    fn main_color(&self) -> RGB8 {
+        self.first_color.into()
+    }
+
+    fn set_main_color(&mut self, color: RGB8) {
+        self.first_color = color.into();
     }
 }