+++ src/sequence/unicolor.rs
@@ -1,6 +1,6 @@
 // led_effects - A collection of LED effects on top of smart-leds.
 // Copyright (C) 2021 Guillaume Cugnet <guillaume@cugnet.eu>
-// Copyright (C) 2021 Jean-Philippe Cugnet <jean-philippe@cugnet.eu>
+// Copyright (C) 2021-2022 Jean-Philippe Cugnet <jean-philippe@cugnet.eu>
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -16,26 +16,34 @@
 
 use smart_leds::RGB8;
 
-use super::{OneParameterSequence, Sequence};
+use super::{ConfigWithMainColor, Sequence};
 
 /// A sequence in which all LEDs have the same color.
 pub struct Unicolor<Color, const N: usize> {
-    /// The color for all LEDs.
-    color: Color,
+    /// The configuration.
+    config: UnicolorConfig<Color>,
     /// The counter.
     counter: usize,
+}
+
+/// The configuration for unicolor sequences.
+#[derive(Debug, Clone, Copy)]
+pub struct UnicolorConfig<Color> {
+    /// The color for all LEDs.
+    pub color: Color,
 }
 
 impl<Color: Copy + Into<RGB8>, const N: usize> Sequence<N>
     for Unicolor<Color, N>
 {
-}
+    type Config = UnicolorConfig<Color>;
 
-impl<Color: Copy + Into<RGB8>, const N: usize> OneParameterSequence<Color, N>
-    for Unicolor<Color, N>
-{
-    fn new(color: Color) -> Self {
-        Self { color, counter: 0 }
+    fn new(config: Self::Config) -> Self {
+        Self { config, counter: 0 }
+    }
+
+    fn config(&self) -> Self::Config {
+        self.config
     }
 }
 
@@ -45,9 +53,21 @@ impl<Color: Copy + Into<RGB8>, const N: usize> Iterator for Unicolor<Color, N> {
     fn next(&mut self) -> Option<Self::Item> {
         if self.counter < N {
             self.counter += 1;
-            Some(self.color.into())
+            Some(self.config.color.into())
         } else {
             None
         }
+    }
+}
+
+impl<Color: Copy + Into<RGB8> + From<RGB8>> ConfigWithMainColor
+    for UnicolorConfig<Color>
+{
+    fn main_color(&self) -> RGB8 {
+        self.color.into()
+    }
+
+    fn set_main_color(&mut self, color: RGB8) {
+        self.color = color.into();
     }
 }